Improving readpfb efficiency using vectorization
The original readpfb function could be more computationally efficient if it read in multiple records simultaneously using vectorization instead of one record at a time. Here is the benchmark for the original and the modified functions reading a PFB file (~585 MB) with NX=12,400, NY=9,900, and NZ=61 ten times:

```
> t1 <- PriorityFlow::readpfb(filename = filename, verbose = F)
> t2 <- readPFB(filename = filename, verbose = F)
> identical(t1, t2)
[1] TRUE

> microbenchmark::microbenchmark(PriorityFlow::readpfb(filename = filename, verbose = F), readPFB(filename = filename, verbose = F), times = 10)
Unit: seconds
                                                    expr        min         lq      mean     median         uq        max neval
 PriorityFlow::readpfb(filename = filename, verbose = F) 261.164690 266.208098 268.06639 267.485137 268.506295 280.608106    10
               readPFB(filename = filename, verbose = F)   4.399969   4.410641   4.56486   4.513397   4.552167   5.184658    10
```
The execution time can be reduced by ~98% with the new implementation.